add abstract remote_file downloader
